requirements.txt
# Core
elasticsearch==8.17.0
python-dotenv==1.0.1
pydantic==2.7.4
rich==13.7.1          # pretty terminal output for demo

# LLM clients
anthropic==0.34.2
openai==1.45.0

# Data generation
faker==26.0.0
numpy==1.26.4

# Embeddings (local fallback — can swap for ELSER on Cloud)
sentence-transformers==3.0.1

# Frontend / demo
streamlit==1.37.0
streamlit-extras==0.4.3

# Utilities
httpx==0.27.0
tenacity==8.5.0       # retry logic for API calls
tqdm==4.66.5          # progress bars during ingest
